wipxxxxx

+++ src/message/handlers.rs
@@ -4,11 +4,11 @@ use crate::message::Message;
 use std::sync::mpsc::Sender;
 
 pub trait MessageHandler<Op> {
-    fn get_code(&self) -> u8;
+    fn get_code(&self) -> u32;
     fn handle(&self, message: &mut Message, sender: Sender<Op>);
 }
 pub struct Handlers<Op> {
-    handlers: HashMap<u8, Box<dyn MessageHandler<Op>>>,
+    handlers: HashMap<u32, Box<dyn MessageHandler<Op>>>,
 }
 
 impl<Op> Default for Handlers<Op> {
@@ -31,7 +31,7 @@ impl<Op> Handlers<Op> {
         self.handlers.insert(handler.get_code(), Box::new(handler));
         self
     }
-    pub fn get_handler(&self, code: u8) -> Option<&dyn MessageHandler<Op>> {
+    pub fn get_handler(&self, code: u32) -> Option<&dyn MessageHandler<Op>> {
         self.handlers.get(&code).map(|v| &**v)
     }
 }
